[REFACTOR] type 폴더위치 수정
